requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2

# LangChain + LangGraph + Groq (LLM) + FastEmbed (local embeddings)
langchain
langchain-core
langchain-community
langchain-groq
langchain-chroma
langchain-text-splitters
langgraph

# Local embeddings (ONNX runtime, no torch)
fastembed

# Misc
tenacity

# Document loaders
docx2txt
pypdf
unstructured

# Vector store
chromadb

# Database (Postgres via Neon)
sqlalchemy
psycopg2-binary
python-dotenv

# Auth
bcrypt
itsdangerous